Add isLoggedIn middleware and add subscription logic

+++ controllers/post.js
@@ -1,9 +1,9 @@
-import Interceptor from "../middlewares/interceptor.js";
+import interceptor from "../middlewares/interceptor.js";
 import Post from "../models/post.js";
 import CustomError from "../utils/customError.js";
 import CustomResponse from "../utils/customResponse.js";
 
-export const getPosts = Interceptor(async (req, res) => {
+export const getPosts = interceptor(async (req, res) => {
   const { skip, limit } = req.query;
   const posts = await Post.find().skip(skip).limit(limit);
   return res
@@ -11,7 +11,7 @@ export const getPosts = Interceptor(async (req, res) => {
     .send(new CustomResponse(200, "Posts fetched successfully!", [], posts));
 });
 
-export const createPost = Interceptor(async (req, res) => {
+export const createPost = interceptor(async (req, res) => {
   if (!req.file) return new CustomError("Please provide an image", 400, res);
 
   const { title, caption, tags } = req.body;
@@ -24,18 +24,22 @@ export const createPost = Interceptor(async (req, res) => {
     caption,
     tags,
     image: req.file.location,
+    owner: req.user?._id,
   });
   return res
     .status(201)
     .send(new CustomResponse(201, "Posts added successfully!", [], post));
 });
 
-export const deletePost = Interceptor(async (req, res, next) => {
+export const deletePost = interceptor(async (req, res, next) => {
   const { id } = req.params;
 
   const post = await Post.findById(id);
 
   if (!post) return new CustomError("Post does not exist", 400, res);
+
+  if (post?.owner?._id !== req.user?._id)
+    return new CustomError("You are not authorized for this task", 401, res);
 
   await Post.findByIdAndDelete(id);
